fix [James]: fix upload image on events
