+++ server-side/controllers/cart.controller.js
@@ -1,4 +1,3 @@
-
 const Cart = require("../models/CartModel");
 const Student = require("../models/StudentModel");
 const mongoose = require("mongoose");
