+++ server-side/routes/events.js
@@ -1,5 +1,7 @@
 const express = require("express");
-const { upload } = require("../util/aws.util");
+const multer = require("multer");
+const multerS3 = require("multer-s3");
+const { S3Client } = require("@aws-sdk/client-s3");
 const {
   admin_authenticate,
   both_authenticate,
@@ -23,6 +25,26 @@ const {
 require("dotenv").config();
 
 const router = express.Router();
+const s3Client = new S3Client({
+  region: process.env.AWS_REGION,
+  credentials: {
+    accessKeyId: process.env.AWS_ACCESS_KEY_ID,
+    secretAccessKey: process.env.AWS_SECRET_ACCESS_KEY,
+  },
+});
+
+const upload = multer({
+  storage: multerS3({
+    s3: s3Client,
+    bucket: process.env.AWS_BUCKET_NAME,
+    metadata: (req, file, cb) => {
+      cb(null, { fieldName: file.fieldname });
+    },
+    key: (req, file, cb) => {
+      cb(null, `event/${Date.now()}_${file.originalname}`);
+    },
+  }),
+});
 
 // POST: Create a new Event
 router.post(
